requirements.txt
beautifulsoup4==4.12.3
python-dotenv==1.0.1
PyYAML==6.0.2
rapidfuzz==3.9.7
requests==2.32.3
pytest==8.3.5
